move api tests to tests package
