+++ server/api/server.go
@@ -30,7 +30,7 @@ import (
 	"github.com/tikv/pd/server"
 )
 
-const apiPrefix = "/pd"
+const APIPrefix = "/pd"
 
 // NewHandler creates a HTTP handler for API.
 func NewHandler(_ context.Context, svr *server.Server) (http.Handler, apiutil.APIServiceGroup, error) {
@@ -38,8 +38,8 @@ func NewHandler(_ context.Context, svr *server.Server) (http.Handler, apiutil.AP
 		Name:   "core",
 		IsCore: true,
 	}
-	prefix := apiPrefix + "/api/v1"
-	r := createRouter(apiPrefix, svr)
+	prefix := APIPrefix + "/api/v1"
+	r := createRouter(APIPrefix, svr)
 	router := mux.NewRouter()
 
 	// Following requests are redirected:
@@ -65,7 +65,7 @@ func NewHandler(_ context.Context, svr *server.Server) (http.Handler, apiutil.AP
 	//	"/schedulers/{name}", http.MethodDelete
 	//  Because the writing of all the config of the scheduling service is in the PD,
 	// 	we should not post and delete the scheduler directly in the scheduling service.
-	router.PathPrefix(apiPrefix).Handler(negroni.New(
+	router.PathPrefix(APIPrefix).Handler(negroni.New(
 		serverapi.NewRuntimeServiceValidator(svr, group),
 		serverapi.NewRedirector(svr,
 			serverapi.MicroserviceRedirectRule(
