+++ server/util.go
@@ -17,7 +17,6 @@ package server
 import (
 	"context"
 	"net/http"
-	"path/filepath"
 	"strings"
 
 	"github.com/coreos/go-semver/semver"
@@ -137,18 +136,4 @@ func combineBuilderServerHTTPService(ctx context.Context, svr *Server, serviceBu
 	apiService.UseHandler(router)
 	userHandlers[pdAPIPrefix] = apiService
 	return userHandlers, nil
-}
-
-func isPathInDirectory(path, directory string) bool {
-	absPath, err := filepath.Abs(path)
-	if err != nil {
-		return false
-	}
-
-	absDir, err := filepath.Abs(directory)
-	if err != nil {
-		return false
-	}
-
-	return strings.HasPrefix(absPath, absDir)
 }
