+++ tests/server/api/min_resolved_ts_test.go
@@ -32,7 +32,9 @@ import (
 	"github.com/tikv/pd/pkg/utils/testutil"
 	"github.com/tikv/pd/pkg/utils/typeutil"
 	"github.com/tikv/pd/server"
+	"github.com/tikv/pd/server/api"
 	"github.com/tikv/pd/server/cluster"
+	"github.com/tikv/pd/tests"
 )
 
 type minResolvedTSTestSuite struct {
@@ -53,10 +55,10 @@ func (suite *minResolvedTSTestSuite) SetupSuite() {
 	cluster.DefaultMinResolvedTSPersistenceInterval = suite.defaultInterval
 	re := suite.Require()
 	suite.svr, suite.cleanup = mustNewServer(re)
-	server.MustWaitLeader(re, []*server.Server{suite.svr})
+	tests.MustWaitLeader(re, []*server.Server{suite.svr})
 
 	addr := suite.svr.GetAddr()
-	suite.url = fmt.Sprintf("%s%s/api/v1/min-resolved-ts", addr, apiPrefix)
+	suite.url = fmt.Sprintf("%s%s/api/v1/min-resolved-ts", addr, api.APIPrefix)
 
 	mustBootstrapCluster(re, suite.svr)
 	suite.storesNum = 3
@@ -76,7 +78,7 @@ func (suite *minResolvedTSTestSuite) TestMinResolvedTS() {
 	re := suite.Require()
 	// case1: default run job
 	interval := suite.svr.GetRaftCluster().GetPDServerConfig().MinResolvedTSPersistenceInterval
-	suite.checkMinResolvedTS(re, &minResolvedTS{
+	suite.checkMinResolvedTS(re, &api.MinResolvedTS{
 		MinResolvedTS:   0,
 		IsRealTime:      true,
 		PersistInterval: interval,
@@ -84,7 +86,7 @@ func (suite *minResolvedTSTestSuite) TestMinResolvedTS() {
 	// case2: stop run job
 	zero := typeutil.Duration{Duration: 0}
 	suite.setMinResolvedTSPersistenceInterval(zero)
-	suite.checkMinResolvedTS(re, &minResolvedTS{
+	suite.checkMinResolvedTS(re, &api.MinResolvedTS{
 		MinResolvedTS:   0,
 		IsRealTime:      false,
 		PersistInterval: zero,
@@ -95,7 +97,7 @@ func (suite *minResolvedTSTestSuite) TestMinResolvedTS() {
 	suite.Eventually(func() bool {
 		return interval == suite.svr.GetRaftCluster().GetPDServerConfig().MinResolvedTSPersistenceInterval
 	}, time.Second*10, time.Millisecond*20)
-	suite.checkMinResolvedTS(re, &minResolvedTS{
+	suite.checkMinResolvedTS(re, &api.MinResolvedTS{
 		MinResolvedTS:   0,
 		IsRealTime:      true,
 		PersistInterval: interval,
@@ -103,7 +105,7 @@ func (suite *minResolvedTSTestSuite) TestMinResolvedTS() {
 	// case4: set min resolved ts
 	ts := uint64(233)
 	suite.setAllStoresMinResolvedTS(ts)
-	suite.checkMinResolvedTS(re, &minResolvedTS{
+	suite.checkMinResolvedTS(re, &api.MinResolvedTS{
 		MinResolvedTS:   ts,
 		IsRealTime:      true,
 		PersistInterval: interval,
@@ -111,13 +113,13 @@ func (suite *minResolvedTSTestSuite) TestMinResolvedTS() {
 	// case5: stop persist and return last persist value when interval is 0
 	interval = typeutil.Duration{Duration: 0}
 	suite.setMinResolvedTSPersistenceInterval(interval)
-	suite.checkMinResolvedTS(re, &minResolvedTS{
+	suite.checkMinResolvedTS(re, &api.MinResolvedTS{
 		MinResolvedTS:   ts,
 		IsRealTime:      false,
 		PersistInterval: interval,
 	})
 	suite.setAllStoresMinResolvedTS(ts)
-	suite.checkMinResolvedTS(re, &minResolvedTS{
+	suite.checkMinResolvedTS(re, &api.MinResolvedTS{
 		MinResolvedTS:   ts, // last persist value
 		IsRealTime:      false,
 		PersistInterval: interval,
@@ -147,7 +149,7 @@ func (suite *minResolvedTSTestSuite) TestMinResolvedTSByStores() {
 
 		testStoresID = append(testStoresID, strconv.Itoa(i))
 	}
-	suite.checkMinResolvedTSByStores(re, &minResolvedTS{
+	suite.checkMinResolvedTSByStores(re, &api.MinResolvedTS{
 		MinResolvedTS:       234,
 		IsRealTime:          true,
 		PersistInterval:     interval,
@@ -156,7 +158,7 @@ func (suite *minResolvedTSTestSuite) TestMinResolvedTSByStores() {
 
 	// set all stores min resolved ts.
 	testStoresIDStr := strings.Join(testStoresID, ",")
-	suite.checkMinResolvedTSByStores(re, &minResolvedTS{
+	suite.checkMinResolvedTSByStores(re, &api.MinResolvedTS{
 		MinResolvedTS:       234,
 		IsRealTime:          true,
 		PersistInterval:     interval,
@@ -167,7 +169,7 @@ func (suite *minResolvedTSTestSuite) TestMinResolvedTSByStores() {
 	testStoresID = testStoresID[:len(testStoresID)-1]
 	testStoresIDStr = strings.Join(testStoresID, ",")
 	delete(testMap, uint64(suite.storesNum))
-	suite.checkMinResolvedTSByStores(re, &minResolvedTS{
+	suite.checkMinResolvedTSByStores(re, &api.MinResolvedTS{
 		MinResolvedTS:       234,
 		IsRealTime:          true,
 		PersistInterval:     interval,
@@ -188,12 +190,12 @@ func (suite *minResolvedTSTestSuite) setAllStoresMinResolvedTS(ts uint64) {
 	}
 }
 
-func (suite *minResolvedTSTestSuite) checkMinResolvedTS(re *require.Assertions, expect *minResolvedTS) {
+func (suite *minResolvedTSTestSuite) checkMinResolvedTS(re *require.Assertions, expect *api.MinResolvedTS) {
 	suite.Eventually(func() bool {
 		res, err := testDialClient.Get(suite.url)
 		re.NoError(err)
 		defer res.Body.Close()
-		listResp := &minResolvedTS{}
+		listResp := &api.MinResolvedTS{}
 		err = apiutil.ReadJSON(res.Body, listResp)
 		re.NoError(err)
 		re.Nil(listResp.StoresMinResolvedTS)
@@ -201,13 +203,13 @@ func (suite *minResolvedTSTestSuite) checkMinResolvedTS(re *require.Assertions, 
 	}, time.Second*10, time.Millisecond*20)
 }
 
-func (suite *minResolvedTSTestSuite) checkMinResolvedTSByStores(re *require.Assertions, expect *minResolvedTS, scope string) {
+func (suite *minResolvedTSTestSuite) checkMinResolvedTSByStores(re *require.Assertions, expect *api.MinResolvedTS, scope string) {
 	suite.Eventually(func() bool {
 		url := fmt.Sprintf("%s?scope=%s", suite.url, scope)
 		res, err := testDialClient.Get(url)
 		re.NoError(err)
 		defer res.Body.Close()
-		listResp := &minResolvedTS{}
+		listResp := &api.MinResolvedTS{}
 		err = apiutil.ReadJSON(res.Body, listResp)
 		re.NoError(err)
 		return reflect.DeepEqual(expect, listResp)
