+++ tests/server/api/testutil.go
@@ -16,14 +16,32 @@ package api
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
 	"path"
+	"sort"
+	"sync"
+	"testing"
 
+	"github.com/pingcap/kvproto/pkg/metapb"
+	"github.com/pingcap/kvproto/pkg/pdpb"
+	"github.com/pingcap/log"
 	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/suite"
 
+	"github.com/tikv/pd/pkg/core"
+	"github.com/tikv/pd/pkg/utils/apiutil"
+	"github.com/tikv/pd/pkg/utils/assertutil"
+	"github.com/tikv/pd/pkg/utils/keypath"
+	"github.com/tikv/pd/pkg/utils/logutil"
+	"github.com/tikv/pd/pkg/utils/testutil"
+	"github.com/tikv/pd/pkg/versioninfo"
+	"github.com/tikv/pd/server"
+	"github.com/tikv/pd/server/api"
+	"github.com/tikv/pd/server/config"
 	"github.com/tikv/pd/tests"
 )
 
@@ -85,4 +103,233 @@ func MustCallSchedulerConfigAPI(
 	data, err = io.ReadAll(resp.Body)
 	re.NoError(err)
 	re.Equal(http.StatusOK, resp.StatusCode, string(data))
+}
+
+var (
+	// testDialClient used to dial http request. only used for test.
+	testDialClient = &http.Client{
+		Transport: &http.Transport{
+			DisableKeepAlives: true,
+		},
+	}
+
+	store = &metapb.Store{
+		Id:        1,
+		Address:   "localhost",
+		NodeState: metapb.NodeState_Serving,
+	}
+	peers = []*metapb.Peer{
+		{
+			Id:      2,
+			StoreId: store.GetId(),
+		},
+	}
+	region = &metapb.Region{
+		Id: 8,
+		RegionEpoch: &metapb.RegionEpoch{
+			ConfVer: 1,
+			Version: 1,
+		},
+		Peers: peers,
+	}
+)
+
+func mustNewServer(re *require.Assertions, opts ...func(cfg *config.Config)) (*server.Server, testutil.CleanupFunc) {
+	_, svrs, cleanup := mustNewCluster(re, 1, opts...)
+	return svrs[0], cleanup
+}
+
+var zapLogOnce sync.Once
+
+func mustNewCluster(re *require.Assertions, num int, opts ...func(cfg *config.Config)) ([]*config.Config, []*server.Server, testutil.CleanupFunc) {
+	ctx, cancel := context.WithCancel(context.Background())
+	svrs := make([]*server.Server, 0, num)
+	cfgs := tests.NewTestMultiConfig(assertutil.CheckerWithNilAssert(re), num)
+
+	ch := make(chan *server.Server, num)
+	for _, cfg := range cfgs {
+		go func(cfg *config.Config) {
+			err := logutil.SetupLogger(&cfg.Log, &cfg.Logger, &cfg.LogProps, cfg.Security.RedactInfoLog)
+			re.NoError(err)
+			zapLogOnce.Do(func() {
+				log.ReplaceGlobals(cfg.Logger, cfg.LogProps)
+			})
+			for _, opt := range opts {
+				opt(cfg)
+			}
+			s, err := server.CreateServer(ctx, cfg, nil, api.NewHandler)
+			re.NoError(err)
+			err = s.Run()
+			re.NoError(err)
+			ch <- s
+		}(cfg)
+	}
+
+	for range num {
+		svr := <-ch
+		svrs = append(svrs, svr)
+	}
+	close(ch)
+	// wait etcd and http servers
+	tests.MustWaitLeader(re, svrs)
+
+	// clean up
+	clean := func() {
+		cancel()
+		for _, s := range svrs {
+			s.Close()
+		}
+		for _, cfg := range cfgs {
+			testutil.CleanServer(cfg.DataDir)
+		}
+	}
+
+	return cfgs, svrs, clean
+}
+
+func mustBootstrapCluster(re *require.Assertions, s *server.Server) {
+	grpcPDClient := testutil.MustNewGrpcClient(re, s.GetAddr())
+	req := &pdpb.BootstrapRequest{
+		Header: testutil.NewRequestHeader(keypath.ClusterID()),
+		Store:  store,
+		Region: region,
+	}
+	resp, err := grpcPDClient.Bootstrap(context.Background(), req)
+	re.NoError(err)
+	re.Equal(pdpb.ErrorType_OK, resp.GetHeader().GetError().GetType())
+}
+
+func mustPutRegion(re *require.Assertions, svr *server.Server, regionID, storeID uint64, start, end []byte, opts ...core.RegionCreateOption) *core.RegionInfo {
+	leader := &metapb.Peer{
+		Id:      regionID,
+		StoreId: storeID,
+	}
+	metaRegion := &metapb.Region{
+		Id:          regionID,
+		StartKey:    start,
+		EndKey:      end,
+		Peers:       []*metapb.Peer{leader},
+		RegionEpoch: &metapb.RegionEpoch{ConfVer: 1, Version: 1},
+	}
+	r := core.NewRegionInfo(metaRegion, leader, opts...)
+	err := svr.GetRaftCluster().HandleRegionHeartbeat(r)
+	re.NoError(err)
+	return r
+}
+
+func mustPutStore(re *require.Assertions, svr *server.Server, id uint64, state metapb.StoreState, nodeState metapb.NodeState, labels []*metapb.StoreLabel) {
+	s := &server.GrpcServer{Server: svr}
+	_, err := s.PutStore(context.Background(), &pdpb.PutStoreRequest{
+		Header: &pdpb.RequestHeader{ClusterId: keypath.ClusterID()},
+		Store: &metapb.Store{
+			Id:        id,
+			Address:   fmt.Sprintf("tikv%d", id),
+			State:     state,
+			NodeState: nodeState,
+			Labels:    labels,
+			Version:   versioninfo.MinSupportedVersion(versioninfo.Version2_0).String(),
+		},
+	})
+	re.NoError(err)
+	if state == metapb.StoreState_Up {
+		_, err = s.StoreHeartbeat(context.Background(), &pdpb.StoreHeartbeatRequest{
+			Header: &pdpb.RequestHeader{ClusterId: keypath.ClusterID()},
+			Stats:  &pdpb.StoreStats{StoreId: id},
+		})
+		re.NoError(err)
+	}
+}
+
+func mustRegionHeartbeat(re *require.Assertions, svr *server.Server, region *core.RegionInfo) {
+	cluster := svr.GetRaftCluster()
+	err := cluster.HandleRegionHeartbeat(region)
+	re.NoError(err)
+}
+
+func mustStoreHeartbeat(re *require.Assertions, svr *server.Server, region *pdpb.StoreHeartbeatRequest) {
+	cluster := svr.GetRaftCluster()
+	err := cluster.HandleStoreHeartbeat(region, &pdpb.StoreHeartbeatResponse{})
+	re.NoError(err)
+}
+
+type serviceTestSuite struct {
+	suite.Suite
+	svr     *server.Server
+	cleanup testutil.CleanupFunc
+}
+
+func TestServiceTestSuite(t *testing.T) {
+	suite.Run(t, new(serviceTestSuite))
+}
+
+func (suite *serviceTestSuite) SetupSuite() {
+	re := suite.Require()
+	suite.svr, suite.cleanup = mustNewServer(re)
+	tests.MustWaitLeader(re, []*server.Server{suite.svr})
+
+	mustBootstrapCluster(re, suite.svr)
+	mustPutStore(re, suite.svr, 1, metapb.StoreState_Up, metapb.NodeState_Serving, nil)
+}
+
+func (suite *serviceTestSuite) TearDownSuite() {
+	suite.cleanup()
+}
+
+func (suite *serviceTestSuite) TestServiceLabels() {
+	re := suite.Require()
+	accessPaths := suite.svr.GetServiceLabels("Profile")
+	re.Len(accessPaths, 1)
+	re.Equal("/pd/api/v1/debug/pprof/profile", accessPaths[0].Path)
+	re.Equal("", accessPaths[0].Method)
+	serviceLabel := suite.svr.GetAPIAccessServiceLabel(
+		apiutil.NewAccessPath("/pd/api/v1/debug/pprof/profile", ""))
+	re.Equal("Profile", serviceLabel)
+	serviceLabel = suite.svr.GetAPIAccessServiceLabel(
+		apiutil.NewAccessPath("/pd/api/v1/debug/pprof/profile", http.MethodGet))
+	re.Equal("Profile", serviceLabel)
+
+	accessPaths = suite.svr.GetServiceLabels("GetSchedulerConfig")
+	re.Len(accessPaths, 1)
+	re.Equal("/pd/api/v1/scheduler-config", accessPaths[0].Path)
+	re.Equal("GET", accessPaths[0].Method)
+	accessPaths = suite.svr.GetServiceLabels("HandleSchedulerConfig")
+	re.Len(accessPaths, 4)
+	re.Equal("/pd/api/v1/scheduler-config", accessPaths[0].Path)
+
+	accessPaths = suite.svr.GetServiceLabels("ResignLeader")
+	re.Len(accessPaths, 1)
+	re.Equal("/pd/api/v1/leader/resign", accessPaths[0].Path)
+	re.Equal(http.MethodPost, accessPaths[0].Method)
+	serviceLabel = suite.svr.GetAPIAccessServiceLabel(
+		apiutil.NewAccessPath("/pd/api/v1/leader/resign", http.MethodPost))
+	re.Equal("ResignLeader", serviceLabel)
+	serviceLabel = suite.svr.GetAPIAccessServiceLabel(
+		apiutil.NewAccessPath("/pd/api/v1/leader/resign", http.MethodGet))
+	re.Equal("", serviceLabel)
+	serviceLabel = suite.svr.GetAPIAccessServiceLabel(
+		apiutil.NewAccessPath("/pd/api/v1/leader/resign", ""))
+	re.Equal("", serviceLabel)
+
+	accessPaths = suite.svr.GetServiceLabels("queryMetric")
+	re.Len(accessPaths, 4)
+	sort.Slice(accessPaths, func(i, j int) bool {
+		if accessPaths[i].Path == accessPaths[j].Path {
+			return accessPaths[i].Method < accessPaths[j].Method
+		}
+		return accessPaths[i].Path < accessPaths[j].Path
+	})
+	re.Equal("/pd/api/v1/metric/query", accessPaths[0].Path)
+	re.Equal(http.MethodGet, accessPaths[0].Method)
+	re.Equal("/pd/api/v1/metric/query", accessPaths[1].Path)
+	re.Equal(http.MethodPost, accessPaths[1].Method)
+	re.Equal("/pd/api/v1/metric/query_range", accessPaths[2].Path)
+	re.Equal(http.MethodGet, accessPaths[2].Method)
+	re.Equal("/pd/api/v1/metric/query_range", accessPaths[3].Path)
+	re.Equal(http.MethodPost, accessPaths[3].Method)
+	serviceLabel = suite.svr.GetAPIAccessServiceLabel(
+		apiutil.NewAccessPath("/pd/api/v1/metric/query", http.MethodPost))
+	re.Equal("queryMetric", serviceLabel)
+	serviceLabel = suite.svr.GetAPIAccessServiceLabel(
+		apiutil.NewAccessPath("/pd/api/v1/metric/query", http.MethodGet))
+	re.Equal("queryMetric", serviceLabel)
 }
