+++ tests/server/api/unsafe_operation_test.go
@@ -26,6 +26,8 @@ import (
 	"github.com/tikv/pd/pkg/unsaferecovery"
 	tu "github.com/tikv/pd/pkg/utils/testutil"
 	"github.com/tikv/pd/server"
+	"github.com/tikv/pd/server/api"
+	"github.com/tikv/pd/tests"
 )
 
 type unsafeOperationTestSuite struct {
@@ -42,10 +44,10 @@ func TestUnsafeOperationTestSuite(t *testing.T) {
 func (suite *unsafeOperationTestSuite) SetupTest() {
 	re := suite.Require()
 	suite.svr, suite.cleanup = mustNewServer(re)
-	server.MustWaitLeader(re, []*server.Server{suite.svr})
+	tests.MustWaitLeader(re, []*server.Server{suite.svr})
 
 	addr := suite.svr.GetAddr()
-	suite.urlPrefix = fmt.Sprintf("%s%s/api/v1/admin/unsafe", addr, apiPrefix)
+	suite.urlPrefix = fmt.Sprintf("%s%s/api/v1/admin/unsafe", addr, api.APIPrefix)
 
 	mustBootstrapCluster(re, suite.svr)
 	mustPutStore(re, suite.svr, 1, metapb.StoreState_Offline, metapb.NodeState_Removing, nil)
